requirements.txt
tensorflow_gpu==1.12.0
torch==1.0.0
scipy==1.0.0
numpy==1.14.2
torchvision==0.2.1
Pillow==6.0.0
tensorboardX==1.6
